extras all working

+++ db/migrate/20170418154014_create_extra_items.rb
@@ -4,7 +4,7 @@ class CreateExtraItems < ActiveRecord::Migration[5.0]
       t.references :customer_plan, index: true, foreign_key: true
       t.references :product, index: true, foreign_key: true
       t.integer :quantity_per_week
-      t.integer :monthly_price
+      t.float :monthly_price
 
       t.timestamps
     end
